requirements.txt
bs4
lxml
pillow
requests
